Rename "home" folder to "homePage" folder
