+++ components/homePage/ArtsAndLifeSection.tsx
