+++ components/homePage/Column.tsx
