+++ components/homePage/DesktopRow.tsx
