+++ components/homePage/LeftSection.tsx
