+++ components/homePage/OpinionSection.tsx
