+++ components/homePage/SectionProps.tsx
