+++ components/homePage/ThumbnailImage.tsx
