+++ components/homePage/ThumbnailImageWithLink.tsx
