+++ components/homePage/getBorderValue.tsx
