+++ pages/index.tsx
@@ -9,19 +9,19 @@ import { getHomeAsync, getHomeMoreAsync, Home, Post } from "../helpers/wpapi";
 import Wrapper from "../components/Wrapper";
 import { CategoryList } from "../components/CategoryList";
 import LoadingView from "../components/Loading";
-import { TopSection } from "../components/home/TopSection";
-import { MainSection } from "../components/home/MainSection";
-import { LeftSection } from "../components/home/LeftSection";
-import { SportsSection } from "../components/home/SportsSection";
-import { GrindSection } from "../components/home/GrindSection";
-import { OpinionSection } from "../components/home/OpinionSection";
-import { ArtsAndLifeSection } from "../components/home/ArtsAndLifeSection";
-import { SponsoredSection } from "../components/home/SponsoredSection";
-import { MultimediaSection } from "../components/home/MultimediaSection";
-import { MoreFromTheDailySection } from "../components/home/MoreFromTheDailySection";
-import { DesktopRow } from "../components/home/DesktopRow";
-import { Column } from "../components/home/Column";
-import { getBorderValue } from "../components/home/getBorderValue";
+import { TopSection } from "../components/homePage/TopSection";
+import { MainSection } from "../components/homePage/MainSection";
+import { LeftSection } from "../components/homePage/LeftSection";
+import { SportsSection } from "../components/homePage/SportsSection";
+import { GrindSection } from "../components/homePage/GrindSection";
+import { OpinionSection } from "../components/homePage/OpinionSection";
+import { ArtsAndLifeSection } from "../components/homePage/ArtsAndLifeSection";
+import { SponsoredSection } from "../components/homePage/SponsoredSection";
+import { MultimediaSection } from "../components/homePage/MultimediaSection";
+import { MoreFromTheDailySection } from "../components/homePage/MoreFromTheDailySection";
+import { DesktopRow } from "../components/homePage/DesktopRow";
+import { Column } from "../components/homePage/Column";
+import { getBorderValue } from "../components/homePage/getBorderValue";
 
 interface IndexProps {
   homePosts?: Home;
